Climb_GUI_React/backend/requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
opencv-python==4.11.0.86
gunicorn==23.0.0
scipy==1.15.1
pillow==11.1.0
numpy==2.2.2
python-dotenv==1.0.1